[11.x] Make floating-types consistent (#48861)
* make floating-types consistent

* fix tests

* fix tests

* fix tests

+++ src/Illuminate/Database/Schema/Grammars/SQLiteGrammar.php
@@ -698,7 +698,7 @@ class SQLiteGrammar extends Grammar
      */
     protected function typeDouble(Fluent $column)
     {
-        return 'float';
+        return 'double';
     }
 
     /**
